app/requirements.txt
Django
psycopg2-binary
gunicorn
flake8
djangorestframework
drf-yasg
django-cors-headers
django-smtp-ssl
djangorestframework-jwt
Babel
celery
django-autoslug
django-filter
